发布1.0.5，增加Optional转换，Properties和String的转换。另外修复原子类型的转换错误，以及java8模块插件加载错误问题

+++ laf-binding-core/src/main/java/com/jd/laf/binding/converter/simple/AtomicIntegerConverter.java
@@ -26,6 +26,6 @@ public class AtomicIntegerConverter extends NumberConverter {
 
     @Override
     public Class<?> targetType() {
-        return Long.class;
+        return AtomicInteger.class;
     }
 }
